Make TempestConf a subclass of SafeConfigParser
TempestConf was needlessly re-implementing or overwriting the behavior of
SafeConfigParser and now the duplication was removed. The check of whether the
value is a string or unicode in `TempestConf.set` has been removed, since
SafeConfigParser already does that. The `TempestConf.is_modified` method was
removed, since it was deceiving (i.e., not actually checking if the key was
modified, only if it is already configured).