Network create setup
This patch is to enable the creation of networks
in case none was configured previously.
It also rename the old create_tempest_network to
check_and_set_tempest_network, since it wasn't creating
anything.
The new create_tempest_network it's now creating the
router, network and subnet, as well as attaching
the subnet to the router.
Also, it receives a list with all users and create a
network for each one.

Change-Id: I38ec49cc06fd97e67d66b993abbeb71c73574dff