Rename --patch argument and fix how it is processed
The --patch is an old name and is being used to accept input from
an installer or deployment. The new name is --deployer-input.
Also, these values are supposed to take precedence over discovery in the
same way as with command line arguments. The doc string and processing of
this parameter were updated to be consistent with that spec.

Change-Id: I7b20042fafacc060bf58329a7b210e4f19f46d4a