Make sure img_dir exists even if image is already in glance
The value of img_dir is used by tempest in a create image test. If this
script sees that the desired image is already in glance, it was not
copying anything to img_dir even if it had no image file. Now it
will check if an image is needed and download it from glance.

Change-Id: I9596266ff5f4df839c657a7dd3a5b9ac0e1ea9ba