waiters.py: raise BackupException defined in tempest/lib
This patch is to fix the issue which described in the bug #1632410

A timeout is being raised on a volume backup test scenario
which could be that because the volume backup is slow or
we tripped a cinder bug. The real issue is that wait loop
for backups in tempest is referencing a non-existent exception class,
"VolumeBackupException".

- https://review.openstack.org/385137

Closes-bug: #1632410
Resolves: rhbz#1455928

(cherry picked from commit ceb2b4684d18e6fa65b500b6a946d4c13cb606f3)

Change-Id: Ia13229d98553b9b60212095fab36c4f05ad2dcca