Fix race in test_networks.py: don't try to get a possibly-deleted network
In compute/admin/test_networks.py, acting as admin, we can list all
networks created by all projects. *But* we can't get (display details)
for one of these networks because by the time we try to GET that network
it could have been deleted.

The fix tries to GET a known/existing-for-sure/permanent network.

Change-Id: If8365dffc84709ad517b40d205729e4a543ceccb