Remove unused arguments from _error_checker() from negative_rest_client
As https://review.openstack.org/#/c/389318/ removes unused arguments
for error_checker() from rest client but not from negative rest client.
Also  tempest.api.compute.flavors.test_flavors_negative is also removed
where negative rest client is used that's why errors related to _error_checker
does not caught in CI, but it when we backported
https://review.openstack.org/#/c/389318/ this review in tempest-13.0.0,
it broke the above tests, That's why it is also needs to be fixed.